trans. ready

+++ product_service/src/service/product-service.ts
@@ -6,6 +6,7 @@ import { AppValidationError } from "../utility/errors";
 import { ProductInput } from "../dto/product-input";
 import { CategoryRepository } from "../repository/category-repository";
 import { ServiceInput } from "../dto/service-input";
+import { AuthUser } from "../utility/auth";
 
 export class ProductService {
   _repository: ProductRepository;
@@ -17,12 +18,30 @@ export class ProductService {
     return ErrorResponse(404, new Error("method not allowed!"));
   }
 
+  async authorisedUser(user_id: number, productId: string) {
+    const product = await this._repository.getProductById(productId);
+    if (!product) return false;
+    return Number(user_id) === Number(product.seller_id);
+  }
+
   async createProduct(event: APIGatewayEvent) {
-    const input = plainToClass(ProductInput, JSON.parse(event.body!));
+    //
+    const token = event.headers.Authorization;
+    const user = await AuthUser(token);
+    if (!user) return ErrorResponse(403, "authorization failed");
+
+    if (user.user_type.toUpperCase() !== "SELLER") {
+      return ErrorResponse(403, "jou need to join as seller to create product");
+    }
+
+    const input = plainToClass(ProductInput, event.body);
     const error = await AppValidationError(input);
     if (error) return ErrorResponse(404, error);
 
-    const data = await this._repository.createProduct(input);
+    const data = await this._repository.createProduct({
+      ...input,
+      seller_id: user.user_id,
+    });
     await new CategoryRepository().addItem({
       id: input.category_id,
       products: [data._id],
@@ -43,7 +62,27 @@ export class ProductService {
     return SuccessResponse(data);
   }
 
+  async getSellerProducts(event: APIGatewayEvent) {
+    const token = event.headers.Authorization;
+    const user = await AuthUser(token);
+    if (!user) return ErrorResponse(403, "authorization failed");
+
+    if (user.user_type.toUpperCase() !== "SELLER") {
+      return ErrorResponse(403, "you need to join as seller to manage product");
+    }
+    const data = await this._repository.getAllSellerProducts(user.user_id);
+
+    return SuccessResponse(data);
+  }
   async editProduct(event: APIGatewayEvent) {
+    const token = event.headers.Authorization;
+    const user = await AuthUser(token);
+    if (!user) return ErrorResponse(403, "authorization failed");
+
+    if (user.user_type.toUpperCase() !== "SELLER") {
+      return ErrorResponse(403, "you need to join as seller to manage product");
+    }
+
     const productId = event.pathParameters?.id;
     if (!productId) return ErrorResponse(403, "please provide product id");
 
@@ -51,19 +90,38 @@ export class ProductService {
     const error = await AppValidationError(input);
     if (error) return ErrorResponse(404, error);
 
+    const isAuthorised = await this.authorisedUser(user.user_id, productId);
+    if (!isAuthorised)
+      return ErrorResponse(403, "you need to join as seller to manage product");
+
     input.id = productId;
     const data = await this._repository.updateProduct(input);
     return SuccessResponse(data);
   }
 
   async deleteProduct(event: APIGatewayEvent) {
+    const token = event.headers.Authorization;
+    const user = await AuthUser(token);
+    if (!user) return ErrorResponse(403, "authorization failed");
+
+    if (user.user_type.toUpperCase() !== "SELLER") {
+      return ErrorResponse(403, "you need to join as seller to manage product");
+    }
+
     const productId = event.pathParameters?.id;
     if (!productId) return ErrorResponse(403, "please provide product id");
+
+    const isAuthorised = await this.authorisedUser(user.user_id, productId);
+    if (!isAuthorised)
+      return ErrorResponse(
+        403,
+        "you are not authorised to delete this product"
+      );
 
     const { category_id, DeleteResult } = await this._repository.deleteProduct(
       productId
     );
-    await new CategoryRepository().addItem({
+    await new CategoryRepository().removeItem({
       id: category_id,
       products: [productId],
     });
